Devices: clean warning after replace of int with size_t in Matrix.

+++ src/devices/realsense2/realsense2Driver.cpp
@@ -859,9 +859,9 @@ bool realsense2Driver::setExtrinsicParam(Matrix &extrinsic, const rs2_extrinsics
 
     extrinsic.eye();
 
-    for (int j=0; j<extrinsic.rows() - 1; j++)
+    for (size_t j=0; j<extrinsic.rows() - 1; j++)
     {
-        for (int i=0; i<extrinsic.cols() - 1; i++)
+        for (size_t i=0; i<extrinsic.cols() - 1; i++)
         {
             extrinsic[j][i] = values.rotation[i + j*extrinsic.cols()];
         }
